Add book hack (WIP)
Does not yet have a GUI for specifying the command and the link text
because it doesn't yet work properly.

For some strange reason, it seems to only work on vanilla servers.
Bukkit servers change the command link to plain text or to a web link if
the link text is a valid URL.

See #486