Improve Wurst Capes
They will now be bound to the UUID instead of the username and will be
hosted on a new server that should allow far more capes than the old
one.

If the new server can't handle the traffic and goes down, it can be
switched back to the old server without requiring another Wurst update.