:rocket: Replace EventListenerList with HashMaps
Why it's faster: Say there are 500 active DeathListeners in the list.
Every time a RenderEvent fires (60+ times per second!), the old
EventManager would have to loop through the 500 DeathListeners and look
for a RenderListener. This would take a long time, even with 0 active
RenderListeners. The new EventManager would only loop through the Set of
RenderListeners, which would finish almost instantly with 0 of them
being active.